pyproject.toml
[tool.poetry]
name = "csauto-web"
version = "2.3.0"
description = "Script to render CSAuto web"
authors = ["NoPlagiarism <37241775+NoPlagiarism@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.26.0"
Jinja2 = "^3.1.3"
loguru = "^0.7.2"
markdown-it-py = {extras = ["linkify"], version = "^3.0.0"}
babel = "^2.14.0"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
